pyproject.toml
[project]
name = "api-autotests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "pytest>=8.3,<9.0",
    "pytest-xdist>=3.6,<4.0",
    "pytest-rerunfailures>=14.0,<15.0",
    "requests>=2.32,<3.0",
    "pydantic[email]>=2.12,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "allure-pytest>=2.13,<3.0",
    "api-client>=1.3.1",
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
    "faker>=40.1.2",
]


[tool.pytest.ini_options]
pythonpath = [".", "src"]

#Linter + formater settings
[tool.ruff]
line-length = 100
